fix: editor colors
