+++ src/editor/ColorsSidebar.js
@@ -6,14 +6,9 @@
  * WordPress dependencies
  */
 import { __ } from '@wordpress/i18n';
-import { Fragment, useEffect } from '@wordpress/element';
+import { Fragment } from '@wordpress/element';
 import { RangeControl } from '@wordpress/components';
 import { ColorPaletteControl } from '@wordpress/block-editor';
-
-/**
- * Internal dependencies
- */
-import { updateEditorColors } from './utils';
 
 const ColorsSidebar = ( {
 	background_color,
@@ -21,37 +16,30 @@ const ColorsSidebar = ( {
 	overlay_opacity,
 	overlay_color,
 	isOverlay,
-} ) => {
-	// On component mount.
-	useEffect(() => {
-		updateEditorColors( background_color );
-	}, [ background_color ]);
-
-	return (
-		<Fragment>
-			<ColorPaletteControl
-				value={ background_color }
-				onChange={ value => onMetaFieldChange( 'background_color', value || '#FFFFFF' ) }
-				label={ __( 'Background Color' ) }
-			/>
-			{ isOverlay && (
-				<Fragment>
-					<ColorPaletteControl
-						value={ overlay_color }
-						onChange={ value => onMetaFieldChange( 'overlay_color', value || '#000000' ) }
-						label={ __( 'Overlay Color' ) }
-					/>
-					<RangeControl
-						label={ __( 'Overlay opacity' ) }
-						value={ overlay_opacity }
-						onChange={ value => onMetaFieldChange( 'overlay_opacity', value ) }
-						min={ 0 }
-						max={ 100 }
-					/>
-				</Fragment>
-			) }
-		</Fragment>
-	);
-};
+} ) => (
+	<Fragment>
+		<ColorPaletteControl
+			value={ background_color }
+			onChange={ value => onMetaFieldChange( 'background_color', value || '#FFFFFF' ) }
+			label={ __( 'Background Color' ) }
+		/>
+		{ isOverlay && (
+			<Fragment>
+				<ColorPaletteControl
+					value={ overlay_color }
+					onChange={ value => onMetaFieldChange( 'overlay_color', value || '#000000' ) }
+					label={ __( 'Overlay Color' ) }
+				/>
+				<RangeControl
+					label={ __( 'Overlay opacity' ) }
+					value={ overlay_opacity }
+					onChange={ value => onMetaFieldChange( 'overlay_opacity', value ) }
+					min={ 0 }
+					max={ 100 }
+				/>
+			</Fragment>
+		) }
+	</Fragment>
+);
 
 export default ColorsSidebar;
