+++ src/editor/index.js
@@ -22,7 +22,7 @@ import SegmentationSidebar from './SegmentationSidebar';
 import DismissSidebar from './DismissSidebar';
 import ColorsSidebar from './ColorsSidebar';
 import Preview from './Preview';
-import withDismissButtonPreview from './withDismissButtonPreview';
+import EditorAdditions from './EditorAdditions';
 import './style.scss';
 
 // Action dispatchers for the sidebar components.
@@ -40,36 +40,18 @@ const mapDispatchToProps = dispatch => {
 	};
 };
 
+const connectData = compose( [
+	withSelect( optionsFieldsSelector ),
+	withDispatch( mapDispatchToProps ),
+] );
+
 // Connect data to components.
-const StatusSidebarWithData = compose( [
-	withSelect( optionsFieldsSelector ),
-	withDispatch( mapDispatchToProps ),
-] )( StatusSidebar );
-
-const SidebarWithData = compose( [
-	withSelect( optionsFieldsSelector ),
-	withDispatch( mapDispatchToProps ),
-] )( Sidebar );
-
-const FrequencySidebarWithData = compose( [
-	withSelect( optionsFieldsSelector ),
-	withDispatch( mapDispatchToProps ),
-] )( FrequencySidebar );
-
-const SegmentationSidebarWithData = compose( [
-	withSelect( optionsFieldsSelector ),
-	withDispatch( mapDispatchToProps ),
-] )( SegmentationSidebar );
-
-const DismissSidebarWithData = compose( [
-	withSelect( optionsFieldsSelector ),
-	withDispatch( mapDispatchToProps ),
-] )( DismissSidebar );
-
-const ColorsSidebarWithData = compose( [
-	withSelect( optionsFieldsSelector ),
-	withDispatch( mapDispatchToProps ),
-] )( ColorsSidebar );
+const StatusSidebarWithData = connectData( StatusSidebar );
+const SidebarWithData = connectData( Sidebar );
+const FrequencySidebarWithData = connectData( FrequencySidebar );
+const SegmentationSidebarWithData = connectData( SegmentationSidebar );
+const DismissSidebarWithData = connectData( DismissSidebar );
+const ColorsSidebarWithData = connectData( ColorsSidebar );
 
 // Register components.
 registerPlugin( 'newspack-popups-status', { render: StatusSidebarWithData } );
@@ -111,14 +93,14 @@ registerPlugin( 'newspack-popups-segmentation', {
 } );
 
 registerPlugin( 'newspack-popups-dismiss', {
-	render: withDismissButtonPreview( () => (
+	render: () => (
 		<PluginDocumentSettingPanel
 			name="popup-dismiss-panel"
 			title={ __( 'Dismiss Button Settings', 'newspack-popups' ) }
 		>
 			<DismissSidebarWithData />
 		</PluginDocumentSettingPanel>
-	) ),
+	),
 	icon: null,
 } );
 
@@ -131,6 +113,11 @@ registerPlugin( 'newspack-popups-colors', {
 			<ColorsSidebarWithData />
 		</PluginDocumentSettingPanel>
 	),
+	icon: null,
+} );
+
+registerPlugin( 'newspack-popups-editor', {
+	render: EditorAdditions,
 	icon: null,
 } );
